Update dependencies
npm audit was complaining about a few security issues. I don't think any
of them would have affected a running app, but let's update everything
anyway.

The main update is react-scripts from 1.1.5 to 2.x. See
https://github.com/facebook/create-react-app/blob/master/CHANGELOG.md#migrating-from-1x-to-203

There's now a requirement for browserslist to be configured. The list I
set is based on Contentful's browser support list at
https://www.contentful.com/faq/about-contentful/#which-browsers-does-contentful-support
which at the time of writing reads:

> - Internet Explorer 11 (except Rich Text)
> - Chrome, Firefox, Opera Latest version
> - Safari Latest minor version (i.e. 9.1.x)
> - Edge Latest stable release
>
> In addition we also support the previous version for an additional six
> months after a new version has been released.

I took this to mean the latest two versions (as shown on caniuse, since
this is how browserslist defines versions, and this takes care of the
"minor" aspect of how Safari is listed in the support list above) of
each of those browsers, plus IE11.

Another major change is that polyfills are no longer included
automatically. So I've included react-app-polyfill and have the app
loading the IE11 version.